feat(MOD-86): change code structure for better naming
